Add info and style About page and update README
This commit adds info to our about page with the technologies used,
the goal of the web application and credits to PedroTech for the
tutorial.

The page was also styled, and the README was updated to contain the
same info in the about page.

TODO: Deploy the page using fly.